requirements.txt
Django~=3.2.3
djangorestframework~=3.12.4
requests~=2.25.1
attrs~=21.2.0
pytest~=6.2.4